Balenciaga e-website
Balenciaga is a fashion house founded by Cristóbal Balenciaga, a Basque designer, born in the Basque Country, Spain. He had a reputation as a couturier of unbending standards and was referred to as "the master of us all" by Christian Dior. His bubble skirts and odd, feminine, yet ultra-modern shapes were trademarks of the house. The house of Balenciaga is now owned by the French multinational company PPR.
This site is under construction by https://sprl.in/VoltaireISHIMWE